Large EnterpriseApplication.java class broken up
