+++ src/main/java/com/example/enterpriseapplication/RecipeManager.java
@@ -1,0 +1,50 @@
+package com.example.enterpriseapplication;
+
+import com.example.enterpriseapplication.dto.Recipe;
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+import com.google.gson.reflect.TypeToken;
+
+import java.io.*;
+import java.lang.reflect.Type;
+import java.util.ArrayList;
+import java.util.List;
+
+public class RecipeManager {
+    private static final String JSON_FILE = "recipes.json";
+    private static final Gson gson = new GsonBuilder().setPrettyPrinting().create();
+
+    private static List<Recipe> readRecipesFromJson() {
+        try (Reader reader = new FileReader(JSON_FILE)) {
+            Type listType = new TypeToken<ArrayList<Recipe>>() {
+            }.getType();
+            return gson.fromJson(reader, listType);
+        } catch (IOException e) {
+            e.printStackTrace();
+            return new ArrayList<>();
+        }
+    }
+
+    private static void displayRecipes(List<Recipe> recipes) {
+        for (Recipe recipe : recipes) {
+            System.out.println(recipe);
+        }
+    }
+
+    private static void saveRecipesToJson(List<Recipe> recipes) {
+        try (Writer writer = new FileWriter(JSON_FILE)) {
+            gson.toJson(recipes, writer);
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+
+    private static Recipe findRecipeByName(List<Recipe> recipes, String name) {
+        for (Recipe recipe : recipes) {
+            if (recipe.getRecipeName().equals(name)) {
+                return recipe;
+            }
+        }
+        return null;
+    }
+}
