+++ src/main/java/com/example/enterpriseapplication/dto/Recipe.java
@@ -4,6 +4,7 @@ package com.example.enterpriseapplication.dto;
 import lombok.Data;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 
 public @Data
@@ -12,4 +13,14 @@ class Recipe {
     private String recipeName;
     private String[] recipeSteps;
     private String[] recipeIngredients;
+
+    @Override
+    public String toString() {
+        return "Recipe{" +
+                "recipeID=" + recipeID +
+                ", recipeName='" + recipeName + '\'' +
+                ", recipeSteps=" + Arrays.toString(recipeSteps) +
+                ", recipeIngredients=" + Arrays.toString(recipeIngredients) +
+                '}';
+    }
 }
